Migration pre 1.12 Android projects
