+++ example/android/app/src/main/AndroidManifest.xml
@@ -15,12 +15,11 @@
 
     <!--android:name="io.flutter.app.FlutterApplication"-->
 
-    <application
-        android:name=".GlobalApplication"
+    <application        
         android:label="flutter_kakao_login_example"
         android:icon="@mipmap/ic_launcher">
         <activity
-            android:name=".MainActivity"
+            android:name="io.flutter.embedding.android.FlutterActivity"
             android:launchMode="singleTop"
             android:theme="@style/LaunchTheme"
             android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection|fontScale|screenLayout|density"
@@ -30,14 +29,14 @@
                  until Flutter renders its first frame. It can be removed if
                  there is no splash screen (such as the default splash screen
                  defined in @style/LaunchTheme). -->
-            <meta-data
-                android:name="io.flutter.app.android.SplashScreenUntilFirstFrame"
-                android:value="true" />
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+        <meta-data
+    android:name="flutterEmbedding"
+    android:value="2" />
         <!-- 3 -->
         <meta-data
             android:name="com.kakao.sdk.AppKey"
