--- example/android/app/src/main/java/com/josephnk/flutterkakaologinexample/GlobalApplication.java
@@ -1,17 +1,0 @@
-package com.josephnk.flutterkakaologinexample;
-
-import io.flutter.app.FlutterApplication;
-import io.flutter.plugin.common.PluginRegistry;
-import io.flutter.plugins.GeneratedPluginRegistrant;
-
-public class GlobalApplication extends FlutterApplication implements PluginRegistry.PluginRegistrantCallback {
-    @Override
-    public void onCreate() {
-        super.onCreate();
-    }
-
-    @Override
-    public void registerWith(PluginRegistry registry) {
-        GeneratedPluginRegistrant.registerWith(registry);
-    }
-}
