+++ example/android/app/src/main/java/com/josephnk/flutterkakaologinexample/MainActivity.java
@@ -11,16 +11,21 @@ import android.util.Log;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
-import io.flutter.app.FlutterActivity;
+import io.flutter.embedding.android.FlutterActivity;
+import io.flutter.embedding.engine.FlutterEngine;
 import io.flutter.plugins.GeneratedPluginRegistrant;
 
 public class MainActivity extends FlutterActivity {
   private static final String LOG_TAG = "KakaoTalkLogin";
 
   @Override
+  public void configureFlutterEngine(FlutterEngine flutterEngine) {
+    GeneratedPluginRegistrant.registerWith(flutterEngine);
+  }
+
+  @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
 
     String hashKey = getKeyHash(this);
     if (hashKey != null) {
